separando matriz de super vetor em 4 e transformando em double. A normalização já está implementada

+++ trabalho2/lib/glcm.h
@@ -1,18 +1,18 @@
 #ifndef _GLCM_H_
 #define _GLCM_H_
 
-void GLCM(int **, int, int, int*);
-void add_to_vector(int, int*, int *);
-int* north(int **, int, int);
-int* south(int **, int, int);
-int* west(int **, int, int);
-int* east(int **, int, int);
-int* northeast(int **, int, int);
-int* northwest(int **, int, int);
-int* southeast(int **, int, int);
-int* south_west(int **, int, int);
-int calc_energy(int **);
-int calc_homogeneity(int **);
-int calc_contrast(int **);
+void GLCM(int **, int, int, double*);
+void add_to_vector(int, double*, double *);
+double* north(int **, int, int);
+double* south(int **, int, int);
+double* west(int **, int, int);
+double* east(int **, int, int);
+double* northeast(int **, int, int);
+double* northwest(int **, int, int);
+double* southeast(int **, int, int);
+double* south_west(int **, int, int);
+double calc_energy(int **);
+double calc_homogeneity(int **);
+double calc_contrast(int **);
 
 #endif
